requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
sqlalchemy==2.0.36
pyyaml==6.0.1
pydantic==2.9.2
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
python-multipart==0.0.9